backend/requirements.txt
fastapi==0.0.3
openai==1.30.1
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
instructor==1.2.6
langchain==0.2.0
langchain-openai==0.1.7